refactor api to account for chainId
