+++ lib/graphqlHelpers.js
@@ -11,7 +11,7 @@ const graphqlReq = axios.create({
 /**
  * Creates multisig record in faunadb
  *
- * @param {object} multisig an object with address (string) and pubkey JSON
+ * @param {object} multisig an object with address (string), pubkey JSON and chainId
  * @return Returns async function that makes a request to the faunadb graphql endpoint
  */
 const createMultisig = async (multisig) => {
@@ -24,9 +24,11 @@ const createMultisig = async (multisig) => {
           createMultisig(data: {
             address: "${multisig.address}"
             pubkeyJSON: ${JSON.stringify(multisig.pubkeyJSON)}
+            chainId: "${multisig.chainId}
           }) {
             _id
             address
+            chainId
           }
         }
       `,
@@ -37,18 +39,20 @@ const createMultisig = async (multisig) => {
 /**
  * Gets multisig pubkey from faundb
  *
- * @param {string} address Must be a multisig address.
+ * @param {string} address A multisig address.
+ * @param {string} chainId The chainId the multisig belongs to.
  * @return Returns async function that makes a request to the faunadb graphql endpoint
  */
-const getMultisig = async (address) => {
+const getMultisig = async (address, chainId) => {
   return graphqlReq({
     method: "POST",
     data: {
       query: `
         query {
-          getMultisig(address: "${address}") {
+          getMultisig(address: "${address}", chainId: "${chainId}",) {
             address
             pubkeyJSON
+            chainId
           }
         }
       `,
