+++ lib/multisigHelpers.js
@@ -8,12 +8,14 @@ import { createMultisigThresholdPubkey, pubkeyToAddress } from "@cosmjs/amino";
  * @param {array} compressedPubkeys Must be an array of compressed Secp256k1 pubkeys (e.g 'A8B5KVhRz1oQuV1dguzFdGBhHrIU/I+R/QfBZcbZFWVG').
  * @param {number} threshold the number of signers required to sign messages from this multisig
  * @param {string} addressPrefix chain based prefix for the address (e.g. 'cosmos')
+ * @param {string} chainId chain-id for the multisig (e.g. 'cosmoshub-4')
  * @return {string} The multisig address.
  */
 const createMultisigFromCompressedSecp256k1Pubkeys = async (
   compressedPubkeys,
   threshold,
   addressPrefix,
+  chainId,
 ) => {
   const pubkeys = compressedPubkeys.map((compressedPubkey) => {
     return {
@@ -28,8 +30,9 @@ const createMultisigFromCompressedSecp256k1Pubkeys = async (
   const multisig = {
     address: multisigAddress,
     pubkeyJSON: JSON.stringify(multisigPubkey),
+    chainId,
   };
-  const res = await axios.post("/api/multisig", multisig);
+  const res = await axios.post(`/api/chain/${chainId}/multisig`, multisig);
   console.log(res.data);
   return res.data.address;
 };
@@ -48,10 +51,11 @@ const getMultisigAccount = async (address, client) => {
   // available from a node. If the multisig was created with this instance
   // of this tool its pubkey will be available in the fauna datastore
   let accountOnChain = await client.getAccount(address);
+  const chainId = await client.getChainId();
 
   if (!accountOnChain || !accountOnChain.pubkey) {
     console.log("No pubkey on chain for: ", address);
-    const res = await axios.get(`/api/multisig/${address}`);
+    const res = await axios.get(`/api/chain/${chainId}/multisig/${address}`);
 
     if (res.status !== 200) {
       throw new Error("Multisig has no pubkey on node, and was not created using this tool.");
