+++ pages/api/chain/[chainId]/multisig/[multisigAddress]/index.js
@@ -1,12 +1,12 @@
-import { getMultisig } from "../../../../lib/graphqlHelpers";
+import { getMultisig } from "../../../../../../lib/graphqlHelpers";
 
 export default async function (req, res) {
   switch (req.method) {
     case "GET":
       try {
-        const { multisigAddress } = req.query;
-        console.log("Function `getMultisig` invoked", multisigAddress);
-        const getRes = await getMultisig(multisigAddress);
+        const { multisigAddress, chainId } = req.query;
+        console.log("Function `getMultisig` invoked", multisigAddress, chainId);
+        const getRes = await getMultisig(multisigAddress, chainId);
         if (!getRes.data.data.getMultisig) {
           res.status(404).send("Multisig not found");
           return;
